refactor: change store name
