+++ src/components/pages/FormPage/FormPage.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import React from 'react';
 import FormPage from './FormPage';
 import { Provider } from 'react-redux';
-import { store } from '../../../store/storeRedux';
+import { store } from '../../../store/store';
 import OrderingForm from '../../ui/OrderingForm/OrderingForm';
 
 describe('FormPage component testing', () => {
