+++ src/components/pages/HomePage/HomePage.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import HomePage from './HomePage';
 import { Provider } from 'react-redux';
-import { store } from '../../../store/storeRedux';
+import { store } from '../../../store/store';
 
 it('Render home page', () => {
   localStorage.removeItem('searchBarValue');
