+++ src/main.tsx
@@ -4,7 +4,7 @@ import { BrowserRouter } from 'react-router-dom';
 import App from './App';
 import './index.scss';
 import { Provider } from 'react-redux';
-import { store } from './store/storeRedux';
+import { store } from './store/store';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
